requirements.txt
pip>=25.1.1
enum34>=1.1.10
attrs>=25.3.0
argparse>=1.4.0
distro>=1.9.0
PySocks>=1.7.1
pillow>=11.2.1
filelock>=3.18.0
discord>=2.3.2
python-dotenv>=1.1.0
requests>=2.32.3
psutil>=7.0.0
yt-dlp>=2025.4.30
Flask>=3.1.0